cread post, delete, put custom hooks and make work
