+++ src/components/features/actionsMenu/MarkersTableActionsMenu.tsx
@@ -1,17 +1,14 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import WarningActionDialog from '@/components/shared/WarningActionDialog'
-import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { getWarehouseManagementApi } from '@/services/generated-api'
 import { SubmitHandler } from 'react-hook-form'
 import { NewMarkerFormData, newMarkerSchema } from '@/schemas/newMarkerSchema'
 import FormDialog from '@/components/shared/FormDialog'
 import NewMarkerForm from '@/components/features/forms/NewMarkerForm'
-import { BodyType } from '@/services/api'
-import { MarkerDto, MarkerFormDto } from '@/services/model'
-import { useSnackbar } from '@/hooks/useSnackbar'
+import { MarkerDto } from '@/services/model'
 import TableActionsMenu from './TableActionsMenu'
-import axios from 'axios'
+import useDeleteMarker from '@/hooks/services/markers/useDeleteMarker'
+import usePutMarker from '@/hooks/services/markers/usePutMarker'
 
 interface MarkersTableActionsMenuProps {
   marker: MarkerDto
@@ -21,7 +18,8 @@ export default function MarkersTableActionsMenu({ marker }: MarkersTableActionsM
   const { t: translate } = useTranslation()
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
   const [selectedOption, setSelectedOption] = React.useState<string | null>(null)
-  const { showSnackbar } = useSnackbar()
+  const mutationDelete = useDeleteMarker(marker.name!)
+  const mutationUpdate = usePutMarker(marker.name!)
 
   const handleClose = () => {
     setSelectedOption(null)
@@ -36,73 +34,10 @@ export default function MarkersTableActionsMenu({ marker }: MarkersTableActionsM
     setSelectedOption(option)
   }
 
-  const queryClient = useQueryClient()
-
-  const mutationDelete = useMutation({
-    mutationFn: (id: number) => getWarehouseManagementApi().deleteApiMarkerDeleteId(id),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['markers'] })
-      showSnackbar({
-        message: translate('newMarker.snackBar.messages.deleteMarker.success', {
-          name: marker.name
-        }),
-        type: 'success'
-      })
-    },
-    onError: (error: unknown) => {
-      if (axios.isAxiosError(error)) {
-        if (error.response && error.response.status === 400) {
-          let location = error.response.data.Message.split(':').shift().split(' ').pop() //example: zones
-          let specificLocation = error.response.data.Message.split(':').pop() // example: zone 1
-          switch (location) {
-            case 'Zones':
-              location = translate('markers.error.zones')
-              break
-            //TODO: ако има други места, където се използва Markers
-          }
-          showSnackbar({
-            message: `${translate('markers.error.errorMessage', { location: location })}: ${specificLocation}`,
-            type: 'error'
-          })
-        } else {
-          showSnackbar({
-            message: translate('newMarker.snackBar.messages.deleteMarker.error'),
-            type: 'error'
-          })
-        }
-      } else {
-        showSnackbar({
-          message: translate('newMarker.snackBar.messages.deleteMarker.error'),
-          type: 'error'
-        })
-      }
-    }
-  })
-
   const onConfirmClick = () => {
     mutationDelete.mutate(marker.id!)
     handleClose()
   }
-
-  const mutationUpdate = useMutation({
-    mutationFn: ({ id, data }: { id: number; data: BodyType<MarkerFormDto> }) =>
-      getWarehouseManagementApi().putApiMarkerEditId(id, data),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['markers'] })
-      showSnackbar({
-        message: translate('newMarker.snackBar.messages.updateMarker.success', {
-          name: marker.name
-        }),
-        type: 'success'
-      })
-    },
-    onError: () => {
-      showSnackbar({
-        message: translate('newMarker.snackBar.messages.updateMarker.error'),
-        type: 'error'
-      })
-    }
-  })
 
   const handleSubmit: SubmitHandler<NewMarkerFormData> = (data) => {
     mutationUpdate.mutate({ id: marker.id!, data: { name: data.markerName } })
